+++ R/simulation.R
@@ -136,27 +136,38 @@ get_charging_rates_distribution <- function(sessions, unit="year", power_interva
 
 
 # Simulate sessions -------------------------------------------------------
-
-#' Estimate sessions energy values
 #'
-#' @param n number of sessions
-#' @param mu means of univariate GMM
-#' @param sigma covariance matrix of univariate GMM
-#' @param log Logical, true if models have logarithmic transformation
+#' Estimate sessions energy values following a log-normal distribution.
+#' The minimum considered value is 1kWh based on real data analysis.
+#'
+#' @param n integer, number of sessions
+#' @param mu numeric, mean of Gaussian distribution
+#' @param sigma numeric, standard deviation of Gaussian distribution.
+#' If unknown, a recommended value is `sd = mu/3`.
+#' @param log logical, true if models have logarithmic transformation
 #'
 #' @return numeric vector
 #' @keywords internal
 #'
-#' @importFrom stats rnorm
+#' @importFrom stats rlnorm
 #'
 estimate_energy <- function(n, mu, sigma, log) {
-  energy_possible <- rnorm(n*2, mu, sigma)
-  energy <- sample(energy_possible[energy_possible > 0.1], n)
-  if (log) {
-    energy <- exp(energy)
-  } else {
-    energy <- abs(energy)
+  valid_energy <- FALSE
+  while (valid_energy != TRUE) {
+    if (log) {
+      energy_possible <- rlnorm(n, meanlog = mu, sdlog = sigma)
+    } else {
+      energy_possible <- rlnorm(n, meanlog = log(mu), sdlog = log(sigma))
+    }
+    energy_possible <- energy_possible[energy_possible > 1]
+    if (length(energy_possible) > 0) {
+      valid_energy <- TRUE
+    }
   }
+  energy <- sample(
+    energy_possible,
+    size = n, replace = TRUE
+  )
   return( energy )
 }
 
@@ -232,23 +243,30 @@ get_estimated_energy <- function(power_vct, energy_models, energy_log) {
 
 #' Estimate sessions connection values
 #'
-#' @param n number of sessions
-#' @param mu means of bivariate GMM
-#' @param sigma covariance matrix of bivariate GMM
-#' @param log Logical, true if models have logarithmic transformation
+#' Estimate sessions connection values following a Multi-variate Guassian
+#' distribution.
+#' The minimum considered value for duration is 30 minutes.
+#'
+#' @param n integer, number of sessions
+#' @param mu numeric vector, means of bivariate GMM
+#' @param sigma numeric matrix, covariance matrix of bivariate GMM
+#' @param log logical, true if models have logarithmic transformation
 #'
 #' @return vector of numeric values
 #' @keywords internal
 #'
 #' @importFrom MASS mvrnorm
+#' @importFrom dplyr slice_sample
 #'
 estimate_connection <- function(n, mu, sigma, log) {
   ev_connections <- as.data.frame(matrix(mvrnorm(n = n, mu = mu, Sigma = sigma), ncol = 2))
   if (log) {
     ev_connections <- exp(ev_connections)
-  } else {
-    ev_connections <- abs(ev_connections)
   }
+  ev_connections <- slice_sample(
+    ev_connections[ev_connections[[1]] > 0 & ev_connections[[2]] > 0.5, ],
+    n = n, replace = TRUE
+  )
   return( ev_connections )
 }